+++ changelog.md
@@ -22,6 +22,7 @@
 - Techniques include "[object Object]" string literal since recent GCS update \#2781
   - You may need to delete the techniques and reimport to apply the fix.
 - Can't modify Advantage points to be negative. \#2844
+- Roll Confirmation dialog shows "Spell" if the OTF has "S:" even if the item is a Skill \#2851
 
 ## Release 0.18.20 06/18/2026